@@ -5,10 +5,10 @@
   +------------------------------------------------------------------------+
   | Phalcon Developer Tools                                                |
   +------------------------------------------------------------------------+
-  | Copyright (c) 2011-2015 Phalcon Team (http://www.phalconphp.com)       |
+  | Copyright (c) 2011-2016 Phalcon Team (https://www.phalconphp.com)      |
   +------------------------------------------------------------------------+
   | This source file is subject to the New BSD License that is bundled     |
-  | with this package in the file docs/LICENSE.txt.                        |
+  | with this package in the file LICENSE.txt.                             |
   |                                                                        |
   | If you did not receive a copy of the license and are unable to         |
   | obtain it through the world-wide-web, please send an email             |
@@ -16,47 +16,30 @@
   +------------------------------------------------------------------------+
   | Authors: Andres Gutierrez <andres@phalconphp.com>                      |
   |          Eduar Carvajal <eduar@phalconphp.com>                         |
+  |          Serghei Iakovlev <serghei@phalconphp.com>                     |
   +------------------------------------------------------------------------+
 */
 
-error_reporting(E_ALL);
-
 use Phalcon\Script;
-use Phalcon\Version;
 use Phalcon\Script\Color;
+use Phalcon\Devtools\Version;
+use Phalcon\Commands\Builtin\Info;
+use Phalcon\Commands\Builtin\Model;
+use Phalcon\Commands\Builtin\Module;
+use Phalcon\Commands\Builtin\Project;
+use Phalcon\Commands\Builtin\Scaffold;
 use Phalcon\Commands\CommandsListener;
-use Phalcon\Loader;
-use Phalcon\Events\Manager as EventsManager;
+use Phalcon\Commands\Builtin\Webtools;
+use Phalcon\Commands\Builtin\AllModels;
+use Phalcon\Commands\Builtin\Migration;
+use Phalcon\Commands\Builtin\Enumerate;
+use Phalcon\Commands\Builtin\Controller;
+use Phalcon\Commands\Builtin\Console;
 use Phalcon\Exception as PhalconException;
+use Phalcon\Events\Manager as EventsManager;
 
 try {
-    if (!extension_loaded('phalcon')) {
-        throw new Exception(
-            "Phalcon extension isn't installed, follow these instructions to install it: " .
-            'https://docs.phalconphp.com/en/latest/reference/install.html'
-        );
-    }
-
-    $loader = new Loader();
-    $loader
-        ->registerDirs(array(__DIR__ . '/scripts/'))
-        ->registerNamespaces(array('Phalcon' => __DIR__ . '/scripts/'))
-        ->register();
-
-    if (file_exists('.phalcon/autoload.php')) {
-        require_once '.phalcon/autoload.php';
-    }
-
-    if (Version::getId() < Script::COMPATIBLE_VERSION) {
-        throw new Exception(
-            sprintf(
-                "Your Phalcon version isn't compatible with Developer Tools, download the latest at: %s",
-                Script::DOC_DOWNLOAD_URL
-            )
-        );
-    }
-
-    defined('TEMPLATE_PATH') || define('TEMPLATE_PATH', __DIR__ . '/templates');
+    require dirname(__FILE__) . '/bootstrap/autoload.php';
 
     $vendor = sprintf('Phalcon DevTools (%s)', Version::get());
     print PHP_EOL . Color::colorize($vendor, Color::FG_GREEN, Color::AT_BOLD) . PHP_EOL . PHP_EOL;
@@ -67,17 +50,21 @@ try {
 
     $script = new Script($eventsManager);
 
-    $commandsToEnable = array(
-        '\Phalcon\Commands\Builtin\Enumerate',
-        '\Phalcon\Commands\Builtin\Controller',
-        '\Phalcon\Commands\Builtin\Module',
-        '\Phalcon\Commands\Builtin\Model',
-        '\Phalcon\Commands\Builtin\AllModels',
-        '\Phalcon\Commands\Builtin\Project',
-        '\Phalcon\Commands\Builtin\Scaffold',
-        '\Phalcon\Commands\Builtin\Migration',
-        '\Phalcon\Commands\Builtin\Webtools'
-    );
+    $commandsToEnable = [
+        Info::class,
+        Enumerate::class,
+        Controller::class,
+        Module::class,
+        Model::class,
+        AllModels::class,
+        Project::class,
+        Scaffold::class,
+        Migration::class,
+        Webtools::class,
+        Console::class,
+    ];
+
+    $script->loadUserScripts();
 
     foreach ($commandsToEnable as $command) {
         $script->attach(new $command($script, $eventsManager));
