@@ -4,20 +4,24 @@
   +------------------------------------------------------------------------+
   | Phalcon Developer Tools                                                |
   +------------------------------------------------------------------------+
-  | Copyright (c) 2011-2016 Phalcon Team (http://www.phalconphp.com)       |
+  | Copyright (c) 2011-2016 Phalcon Team (https://www.phalconphp.com)      |
   +------------------------------------------------------------------------+
   | This source file is subject to the New BSD License that is bundled     |
-  | with this package in the file docs/LICENSE.txt.                        |
+  | with this package in the file LICENSE.txt.                             |
   |                                                                        |
   | If you did not receive a copy of the license and are unable to         |
   | obtain it through the world-wide-web, please send an email             |
   | to license@phalconphp.com so we can send you a copy immediately.       |
   +------------------------------------------------------------------------+
-  | Authors: Serghei Iakovlev <serghei@phalconphp.com>                     |
+  | Authors: Andres Gutierrez <andres@phalconphp.com>                      |
+  |          Eduar Carvajal <eduar@phalconphp.com>                         |
+  |          Serghei Iakovlev <serghei@phalconphp.com>                     |
   +------------------------------------------------------------------------+
 */
 
 namespace Phalcon\Builder;
+
+use SplFileInfo;
 
 /**
  * Module Builder
@@ -28,18 +32,18 @@ namespace Phalcon\Builder;
  */
 class Module extends Component
 {
-    protected $moduleDirectories = array(
+    protected $moduleDirectories = [
         'config',
         'controllers',
         'models',
         'views',
-    );
+    ];
 
     /**
      * Stores variable values depending on parameters
      * @var array
      */
-    protected $variableValues = array();
+    protected $variableValues = [];
 
     /**
      * Create Builder object
@@ -88,7 +92,7 @@ class Module extends Component
             $modulesDir = $this->path->getRootPath($modulesDir);
         }
 
-        $this->options->offsetSet('modulesDir', realpath($modulesDir));
+        $this->options->offsetSet('modulesDir', $modulesDir);
         $this->options->offsetSet('templatePath', realpath($templatePath));
         $this->options->offsetSet('projectPath', $this->path->getRootPath());
 
@@ -98,7 +102,7 @@ class Module extends Component
             ->createConfig()
             ->createModule();
 
-        $this->_notifySuccess(sprintf(
+        $this->notifySuccess(sprintf(
             'Module "%s" was successfully created.',
             $this->options->get('name')
         ));
@@ -122,10 +126,38 @@ class Module extends Component
             ));
         }
 
-        mkdir($modulesDir . DIRECTORY_SEPARATOR . $moduleName, 0777, true);
+        $modulesPath = new SplFileInfo($modulesDir);
+        $modulePath  = $modulesDir. DIRECTORY_SEPARATOR . $moduleName;
 
-        foreach ($this->moduleDirectories as $dir) {
-            mkdir($modulesDir . DIRECTORY_SEPARATOR . $moduleName . DIRECTORY_SEPARATOR . $dir, 0777, true);
+        try {
+            if ($modulesPath->isFile() && !$modulesPath->isDir()) {
+                throw new BuilderException(
+                    sprintf(
+                        "Builder expects a directory for 'modulesDir'. But %s is a file.",
+                        $modulesPath->getPathname()
+                    )
+                );
+            } elseif ($modulesPath->isReadable() && !mkdir($modulePath, 0777, true)) {
+                throw new BuilderException("Unable to create module directory. Check permissions.");
+            }
+
+            foreach ($this->moduleDirectories as $dir) {
+                $path = $modulePath . DIRECTORY_SEPARATOR . $dir;
+                if (!mkdir($path, 0777, true)) {
+                    throw new BuilderException(
+                        sprintf(
+                            "Unable to create %s directory. Check permissions.",
+                            $path
+                        )
+                    );
+                }
+            }
+        } catch (\Exception $e) {
+            throw new BuilderException(
+                $e->getMessage(),
+                $e->getCode(),
+                ($e instanceof BuilderException ? null : $e)
+            );
         }
 
         return $this;
@@ -198,7 +230,7 @@ class Module extends Component
      */
     protected function getVariableValues()
     {
-        $variableValuesResult = array();
+        $variableValuesResult = [];
         $variablesJsonFile = $this->options->get('templatePath') . DIRECTORY_SEPARATOR . 'variables.json';
 
         if (file_exists($variablesJsonFile)) {
