@@ -4,16 +4,18 @@
   +------------------------------------------------------------------------+
   | Phalcon Developer Tools                                                |
   +------------------------------------------------------------------------+
-  | Copyright (c) 2011-2016 Phalcon Team (http://www.phalconphp.com)       |
+  | Copyright (c) 2011-2016 Phalcon Team (https://www.phalconphp.com)      |
   +------------------------------------------------------------------------+
   | This source file is subject to the New BSD License that is bundled     |
-  | with this package in the file docs/LICENSE.txt.                        |
+  | with this package in the file LICENSE.txt.                             |
   |                                                                        |
   | If you did not receive a copy of the license and are unable to         |
   | obtain it through the world-wide-web, please send an email             |
   | to license@phalconphp.com so we can send you a copy immediately.       |
   +------------------------------------------------------------------------+
-  | Authors: Serghei Iakovlev <serghei@phalconphp.com>                     |
+  | Authors: Andres Gutierrez <andres@phalconphp.com>                      |
+  |          Eduar Carvajal <eduar@phalconphp.com>                         |
+  |          Serghei Iakovlev <serghei@phalconphp.com>                     |
   +------------------------------------------------------------------------+
 */
 
@@ -47,10 +49,10 @@ class Path
      */
     public function getConfig($type = null)
     {
-        $types = array('php' => true, 'ini' => true);
+        $types = ['php' => true, 'ini' => true];
         $type  = isset($types[$type]) ? $type : 'ini';
 
-        foreach (array('app/config/', 'config/', 'apps/config/', 'apps/frontend/config/') as $configPath) {
+        foreach (['app/config/', 'config/', 'apps/config/', 'apps/frontend/config/'] as $configPath) {
             if ('ini' == $type && file_exists($this->rootPath . $configPath . 'config.ini')) {
                 return new ConfigIni($this->rootPath . $configPath . 'config.ini');
             } else {
@@ -87,7 +89,7 @@ class Path
 
     public function setRootPath($path)
     {
-        $this->rootPath = rtrim($path, '\\/') . DIRECTORY_SEPARATOR;
+        $this->rootPath = rtrim(str_replace('/', DIRECTORY_SEPARATOR, $path), '\\/') . DIRECTORY_SEPARATOR;
 
         return $this;
     }
