@@ -4,10 +4,10 @@
   +------------------------------------------------------------------------+
   | Phalcon Developer Tools                                                |
   +------------------------------------------------------------------------+
-  | Copyright (c) 2011-2016 Phalcon Team (http://www.phalconphp.com)       |
+  | Copyright (c) 2011-2016 Phalcon Team (https://www.phalconphp.com)      |
   +------------------------------------------------------------------------+
   | This source file is subject to the New BSD License that is bundled     |
-  | with this package in the file docs/LICENSE.txt.                        |
+  | with this package in the file LICENSE.txt.                             |
   |                                                                        |
   | If you did not receive a copy of the license and are unable to         |
   | obtain it through the world-wide-web, please send an email             |
@@ -19,6 +19,11 @@
 */
 
 namespace Phalcon\Builder;
+
+use Phalcon\Builder\Project\Cli;
+use Phalcon\Builder\Project\Micro;
+use Phalcon\Builder\Project\Simple;
+use Phalcon\Builder\Project\Modules;
 
 /**
  * Project Builder
@@ -44,12 +49,12 @@ class Project extends Component
      * Available Project Types
      * @var array
      */
-    private $_types = array(
-        self::TYPE_MICRO   => '\Phalcon\Builder\Project\Micro',
-        self::TYPE_SIMPLE  => '\Phalcon\Builder\Project\Simple',
-        self::TYPE_MODULES => '\Phalcon\Builder\Project\Modules',
-        self::TYPE_CLI     => '\Phalcon\Builder\Project\Cli',
-    );
+    private $_types = [
+        self::TYPE_MICRO   => Micro::class,
+        self::TYPE_SIMPLE  => Simple::class,
+        self::TYPE_MODULES => Modules::class,
+        self::TYPE_CLI     => Cli::class,
+    ];
 
     /**
      * Project build
@@ -109,7 +114,7 @@ class Project extends Component
         $success = $builder->build();
 
         if ($success === true) {
-            $this->_notifySuccess(sprintf(
+            $this->notifySuccess(sprintf(
                 'Project "%s" was successfully created.',
                 $this->options->get('name')
             ));
