@@ -4,10 +4,10 @@
   +------------------------------------------------------------------------+
   | Phalcon Developer Tools                                                |
   +------------------------------------------------------------------------+
-  | Copyright (c) 2011-2016 Phalcon Team (http://www.phalconphp.com)       |
+  | Copyright (c) 2011-2016 Phalcon Team (https://www.phalconphp.com)      |
   +------------------------------------------------------------------------+
   | This source file is subject to the New BSD License that is bundled     |
-  | with this package in the file docs/LICENSE.txt.                        |
+  | with this package in the file LICENSE.txt.                             |
   |                                                                        |
   | If you did not receive a copy of the license and are unable to         |
   | obtain it through the world-wide-web, please send an email             |
@@ -36,12 +36,13 @@ class Cli extends ProjectBuilder
      * Project directories
      * @var array
      */
-    protected $projectDirectories = array(
+    protected $projectDirectories = [
         'app',
         'app/config',
         'app/tasks',
+        'app/models',
         '.phalcon',
-    );
+    ];
 
     /**
      * Creates the configuration
@@ -74,8 +75,8 @@ class Cli extends ProjectBuilder
      */
     private function createBootstrapFiles()
     {
-        $getFile = $this->options->get('templatePath') . '/project/cli/cli.php';
-        $putFile = $this->options->get('projectPath') . 'app/cli.php';
+        $getFile = $this->options->get('templatePath') . '/project/cli/bootstrap.php';
+        $putFile = $this->options->get('projectPath') . 'app/bootstrap.php';
         $this->generateFile($getFile, $putFile);
 
         return $this;
