@@ -4,10 +4,10 @@
   +------------------------------------------------------------------------+
   | Phalcon Developer Tools                                                |
   +------------------------------------------------------------------------+
-  | Copyright (c) 2011-2016 Phalcon Team (http://www.phalconphp.com)       |
+  | Copyright (c) 2011-2016 Phalcon Team (https://www.phalconphp.com)      |
   +------------------------------------------------------------------------+
   | This source file is subject to the New BSD License that is bundled     |
-  | with this package in the file docs/LICENSE.txt.                        |
+  | with this package in the file LICENSE.txt.                             |
   |                                                                        |
   | If you did not receive a copy of the license and are unable to         |
   | obtain it through the world-wide-web, please send an email             |
@@ -15,6 +15,7 @@
   +------------------------------------------------------------------------+
   | Authors: Andres Gutierrez <andres@phalconphp.com>                      |
   |          Eduar Carvajal <eduar@phalconphp.com>                         |
+  |          Serghei Iakovlev <serghei@phalconphp.com>                     |
   +------------------------------------------------------------------------+
 */
 
@@ -35,7 +36,7 @@ abstract class ProjectBuilder
      * Stores variable values depending on parameters
      * @var array
      */
-    protected $variableValues = array();
+    protected $variableValues = [];
 
     /**
      * Builder options
@@ -47,7 +48,7 @@ abstract class ProjectBuilder
      * Project directories
      * @var array
      */
-    protected $projectDirectories = array();
+    protected $projectDirectories = [];
 
     public function __construct(Options $options)
     {
@@ -80,7 +81,7 @@ abstract class ProjectBuilder
      */
     protected function getVariableValues()
     {
-        $variableValuesResult = array();
+        $variableValuesResult = [];
         $variablesJsonFile =
             $this->options->get('templatePath') . DIRECTORY_SEPARATOR
             . 'project' . DIRECTORY_SEPARATOR
