@@ -4,19 +4,20 @@
   +------------------------------------------------------------------------+
   | Phalcon Developer Tools                                                |
   +------------------------------------------------------------------------+
-  | Copyright (c) 2011-2016 Phalcon Team (http://www.phalconphp.com)       |
+  | Copyright (c) 2011-2016 Phalcon Team (https://www.phalconphp.com)      |
   +------------------------------------------------------------------------+
   | This source file is subject to the New BSD License that is bundled     |
-  | with this package in the file docs/LICENSE.txt.                        |
+  | with this package in the file LICENSE.txt.                             |
   |                                                                        |
   | If you did not receive a copy of the license and are unable to         |
   | obtain it through the world-wide-web, please send an email             |
   | to license@phalconphp.com so we can send you a copy immediately.       |
   +------------------------------------------------------------------------+
-  | Authors: Serghei Iakovlev <serghei@phalconphp.com>                     |
+  | Authors: Andres Gutierrez <andres@phalconphp.com>                      |
+  |          Eduar Carvajal <eduar@phalconphp.com>                         |
+  |          Serghei Iakovlev <serghei@phalconphp.com>                     |
   +------------------------------------------------------------------------+
 */
-
 namespace Phalcon\Builder;
 
 use Phalcon\Config;
@@ -51,7 +52,7 @@ class Options extends Config
      */
     public function has($key)
     {
-        return (isset($this->key) || array_key_exists($key, $this->toArray()));
+        return (isset($this->{$key}) || array_key_exists($key, $this->toArray()));
     }
 
     /**
@@ -65,7 +66,7 @@ class Options extends Config
      */
     public function contains($key)
     {
-        return $this->has($key) && $this->$key;
+        return $this->has($key) && $this->{$key};
     }
 
     /**
@@ -78,6 +79,6 @@ class Options extends Config
      */
     public function get($key, $default = null)
     {
-        return $this->contains($key) ? $this->$key : $default;
+        return $this->contains($key) ? $this->{$key} : $default;
     }
 }
