@@ -4,10 +4,10 @@
   +------------------------------------------------------------------------+
   | Phalcon Developer Tools                                                |
   +------------------------------------------------------------------------+
-  | Copyright (c) 2011-2016 Phalcon Team (http://www.phalconphp.com)       |
+  | Copyright (c) 2011-2016 Phalcon Team (https://www.phalconphp.com)      |
   +------------------------------------------------------------------------+
   | This source file is subject to the New BSD License that is bundled     |
-  | with this package in the file docs/LICENSE.txt.                        |
+  | with this package in the file LICENSE.txt.                             |
   |                                                                        |
   | If you did not receive a copy of the license and are unable to         |
   | obtain it through the world-wide-web, please send an email             |
@@ -43,7 +43,7 @@ class AllModels extends Command
      */
     public function getPossibleParams()
     {
-        return array(
+        return [
             'config=s'    => 'Configuration file [optional]',
             'models=s'    => 'Models directory [optional]',
             'schema=s'    => 'Name of the schema. [optional]',
@@ -57,8 +57,9 @@ class AllModels extends Command
             'validations' => 'Define possible domain validation according to conventions',
             'directory=s' => 'Base path on which project will be created [optional]',
             'mapcolumn'   => 'Get some code for map columns [optional]',
-            'abstract'    => 'Abstract Model [optional]'
-        );
+            'abstract'    => 'Abstract Model [optional]',
+            'help'        => 'Shows this help [optional]',
+        ];
     }
 
     /**
@@ -112,7 +113,7 @@ class AllModels extends Command
             $modelsDir = $this->path->getRootPath($modelsDir);
         }
 
-        $modelBuilder = new AllModelsBuilder(array(
+        $modelBuilder = new AllModelsBuilder([
             'force' => $this->isReceivedOption('force'),
             'config' => $config,
             'schema' => $this->getOption('schema'),
@@ -126,7 +127,7 @@ class AllModels extends Command
             'modelsDir' => $modelsDir,
             'mapColumn' => $this->isReceivedOption('mapcolumn'),
             'abstract' => $this->isReceivedOption('abstract')
-        ));
+        ]);
 
         $modelBuilder->build();
     }
@@ -138,7 +139,7 @@ class AllModels extends Command
      */
     public function getCommands()
     {
-        return array('all-models', 'create-all-models');
+        return ['all-models', 'create-all-models'];
     }
 
     /**
