@@ -4,10 +4,10 @@
   +------------------------------------------------------------------------+
   | Phalcon Developer Tools                                                |
   +------------------------------------------------------------------------+
-  | Copyright (c) 2011-2016 Phalcon Team (http://www.phalconphp.com)       |
+  | Copyright (c) 2011-2016 Phalcon Team (https://www.phalconphp.com)      |
   +------------------------------------------------------------------------+
   | This source file is subject to the New BSD License that is bundled     |
-  | with this package in the file docs/LICENSE.txt.                        |
+  | with this package in the file LICENSE.txt.                             |
   |                                                                        |
   | If you did not receive a copy of the license and are unable to         |
   | obtain it through the world-wide-web, please send an email             |
@@ -39,7 +39,9 @@ class Enumerate extends Command
      */
     public function getPossibleParams()
     {
-        return array();
+        return [
+            'help' => 'Shows this help [optional]',
+        ];
     }
 
     /**
@@ -73,7 +75,7 @@ class Enumerate extends Command
      */
     public function getCommands()
     {
-        return array('commands', 'list', 'enumerate');
+        return ['commands', 'list', 'enumerate'];
     }
 
     /**
@@ -94,9 +96,9 @@ class Enumerate extends Command
     public function getHelp()
     {
         print Color::head('Help:') . PHP_EOL;
-        print Color::colorize('  Lists the commands available in Phalcon devtools') . PHP_EOL . PHP_EOL;
+        print Color::colorize('  Lists the commands available in Phalcon DevTools') . PHP_EOL . PHP_EOL;
 
-        $this->run(array());
+        $this->run([]);
     }
 
     /**
