@@ -4,10 +4,10 @@
   +------------------------------------------------------------------------+
   | Phalcon Developer Tools                                                |
   +------------------------------------------------------------------------+
-  | Copyright (c) 2011-2016 Phalcon Team (http://www.phalconphp.com)       |
+  | Copyright (c) 2011-2016 Phalcon Team (https://www.phalconphp.com)      |
   +------------------------------------------------------------------------+
   | This source file is subject to the New BSD License that is bundled     |
-  | with this package in the file docs/LICENSE.txt.                        |
+  | with this package in the file LICENSE.txt.                             |
   |                                                                        |
   | If you did not receive a copy of the license and are unable to         |
   | obtain it through the world-wide-web, please send an email             |
@@ -40,23 +40,23 @@ class CommandsListener
      */
     public function beforeCommand(Event $event, Command $command)
     {
+        $parameters = $command->parseParameters([], ['h' => 'help']);
+
+        if (
+            count($parameters) < ($command->getRequiredParams() + 1) ||
+            $command->isReceivedOption(['help', 'h', '?']) ||
+            in_array($command->getOption(1), ['help', 'h', '?'])
+        ) {
+            $command->getHelp();
+
+            return false;
+        }
+
         if ($command->canBeExternal() == false) {
             $path = $command->getOption('directory');
             if (!file_exists($path.'.phalcon') || !is_dir($path.'.phalcon')) {
                 throw new CommandsException('This command should be invoked inside a Phalcon project directory.');
             }
-        }
-
-        $parameters = $command->parseParameters();
-
-        if (
-            count($parameters) < ($command->getRequiredParams() + 1) ||
-            $command->isReceivedOption(['help', '?']) ||
-            in_array($command->getOption(1), ['help', '?'])
-        ) {
-            $command->getHelp();
-
-            return false;
         }
 
         return true;
