@@ -4,10 +4,10 @@
   +------------------------------------------------------------------------+
   | Phalcon Developer Tools                                                |
   +------------------------------------------------------------------------+
-  | Copyright (c) 2011-2016 Phalcon Team (http://www.phalconphp.com)       |
+  | Copyright (c) 2011-2016 Phalcon Team (https://www.phalconphp.com)      |
   +------------------------------------------------------------------------+
   | This source file is subject to the New BSD License that is bundled     |
-  | with this package in the file docs/LICENSE.txt.                        |
+  | with this package in the file LICENSE.txt.                             |
   |                                                                        |
   | If you did not receive a copy of the license and are unable to         |
   | obtain it through the world-wide-web, please send an email             |
@@ -21,13 +21,19 @@
 
 namespace Phalcon;
 
-use Phalcon\Script\Color;
-use Phalcon\Version\Item as VersionItem;
-use Phalcon\Mvc\Model\Migration as ModelMigration;
-use Phalcon\Mvc\Model\Exception as ModelException;
-use Phalcon\Db\Exception as DbException;
-use Phalcon\Script\ScriptException;
+use Phalcon\Db\Index;
 use DirectoryIterator;
+use Phalcon\Db\Column;
+use Phalcon\Db\Adapter;
+use Phalcon\Script\Color;
+use Phalcon\Db\AdapterInterface;
+use Phalcon\Version\ItemInterface;
+use Phalcon\Script\ScriptException;
+use Phalcon\Db\Exception as DbException;
+use Phalcon\Mvc\Model\Exception as ModelException;
+use Phalcon\Mvc\Model\Migration as ModelMigration;
+use Phalcon\Version\IncrementalItem as IncrementalVersion;
+use Phalcon\Version\ItemCollection as VersionCollection;
 
 /**
  * Migrations Class
@@ -37,94 +43,15 @@ use DirectoryIterator;
 class Migrations
 {
     /**
-     * Migration database connection
-     * @var \Phalcon\Db\AdapterInterface
+     * name of the migration table
      */
-    protected static $_connection;
+    const MIGRATION_LOG_TABLE = 'phalcon_migrations';
 
     /**
-     * Configuration holder
-     * @var \Phalcon\Config
+     * Filename or db connection to store migrations log
+     * @var mixed|Adapter\Pdo
      */
-    protected static $_config;
-
-    /**
-     * Migration version file name
-     * @var string
-     */
-    protected static $_migrationFid;
-
-    /**
-     * Generate migrations
-     *
-     * @param array $options
-     *
-     * @throws \Exception
-     * @todo Refactor so full migrations history is kept in log
-     */
-    public static function generate(array $options)
-    {
-        $tableName = $options['tableName'];
-        $exportData = $options['exportData'];
-        $migrationsDir = $options['migrationsDir'];
-        $originalVersion = $options['originalVersion'];
-        $force = $options['force'];
-        $config = $options['config'];
-
-        if ($migrationsDir && !file_exists($migrationsDir)) {
-            mkdir($migrationsDir, 0777, true);
-        }
-
-        if ($originalVersion) {
-            if (!preg_match('/[a-z0-9](\.[a-z0-9]+)*/', $originalVersion, $matches)) {
-                throw new \Exception("Version {$originalVersion} is invalid");
-            }
-
-            $originalVersion = $matches[0];
-            $version = new VersionItem($originalVersion, 3);
-            if (file_exists($migrationsDir . DIRECTORY_SEPARATOR . $version) && !$force) {
-                throw new \Exception("Version {$version} is already generated");
-            }
-        } else {
-            $versions = ModelMigration::scanForVersions($migrationsDir);
-
-            if (!count($versions)) {
-                $version = new VersionItem('1.0.0');
-            } else {
-                $version = VersionItem::maximum($versions);
-                $version = $version->addMinor(1);
-            }
-        }
-
-        if (!file_exists($migrationsDir . DIRECTORY_SEPARATOR . $version)) {
-            mkdir($migrationsDir . DIRECTORY_SEPARATOR . $version);
-        }
-
-        if (!isset($config->database)) {
-            throw new \Exception("Cannot load database configuration");
-        }
-
-        ModelMigration::setup($config->database);
-
-        ModelMigration::setSkipAutoIncrement($options['no-ai']);
-        ModelMigration::setMigrationPath($migrationsDir);
-
-        if ($tableName == 'all') {
-            $migrations = ModelMigration::generateAll($version, $exportData);
-            foreach ($migrations as $tableName => $migration) {
-                file_put_contents($migrationsDir . '/' . $version . '/' . $tableName . '.php',
-                    '<?php ' . PHP_EOL . PHP_EOL . $migration);
-            }
-        } else {
-            $migration = ModelMigration::generate($version, $tableName, $exportData);
-            file_put_contents($migrationsDir . '/' . $version . '/' . $tableName . '.php',
-                '<?php ' . PHP_EOL . PHP_EOL . $migration);
-        }
-
-        if (self::isConsole()) {
-            print Color::success('Version ' . $version . ' was successfully generated') . PHP_EOL;
-        }
-    }
+    protected static $_storage;
 
     /**
      * Check if the script is running on Console mode
@@ -133,7 +60,111 @@ class Migrations
      */
     public static function isConsole()
     {
-        return PHP_SAPI == 'cli';
+        return PHP_SAPI === 'cli';
+    }
+
+    /**
+     * Generate migrations
+     *
+     * @param array $options
+     *
+     * @throws \Exception
+     * @throws \LogicException
+     * @throws \RuntimeException
+     */
+    public static function generate(array $options)
+    {
+        $tableName = $options['tableName'];
+        $exportData = $options['exportData'];
+        $migrationsDir = $options['migrationsDir'];
+        $version = isset($options['version']) ? $options['version'] : null;
+        $force = $options['force'];
+        $config = $options['config'];
+        $descr = isset($options['descr']) ? $options['descr'] : null;
+        $noAutoIncrement = isset($options['noAutoIncrement']) ? $options['noAutoIncrement'] : null;
+
+        // Migrations directory
+        if ($migrationsDir && !file_exists($migrationsDir)) {
+            mkdir($migrationsDir, 0755, true);
+        }
+
+        // Use timestamped version if description is provided
+        if ($descr) {
+            $version = (string)(int)(microtime(true) * pow(10, 6));
+            VersionCollection::setType(VersionCollection::TYPE_TIMESTAMPED);
+            $versionItem = VersionCollection::createItem($version . '_' . $descr);
+
+            // Elsewhere use old-style incremental versioning
+            // The version is specified
+        } elseif ($version) {
+            VersionCollection::setType(VersionCollection::TYPE_INCREMENTAL);
+            $versionItem = VersionCollection::createItem($version);
+
+            // The version is guessed automatically
+        } else {
+            VersionCollection::setType(VersionCollection::TYPE_INCREMENTAL);
+            $versionItems = ModelMigration::scanForVersions($migrationsDir);
+
+            if (!isset($versionItems[0])) {
+                $versionItem = VersionCollection::createItem('1.0.0');
+
+            } else {
+                /** @var IncrementalVersion $versionItem */
+                $versionItem = VersionCollection::maximum($versionItems);
+                $versionItem = $versionItem->addMinor(1);
+            }
+        }
+
+        // Path to migration dir
+        $migrationPath = rtrim($migrationsDir, '\\/') . DIRECTORY_SEPARATOR . $versionItem->getVersion();
+        if (!file_exists($migrationPath)) {
+            if (is_writable(dirname($migrationPath))) {
+                mkdir($migrationPath);
+            } else {
+                throw new \RuntimeException("Unable to write '{$migrationPath}' directory. Permission denied");
+            }
+        } elseif (!$force) {
+            throw new \LogicException('Version ' . $versionItem->getVersion() . ' already exists');
+        }
+
+        // Try to connect to the DB
+        if (!isset($config->database)) {
+            throw new \RuntimeException('Cannot load database configuration');
+        }
+        ModelMigration::setup($config->database);
+        ModelMigration::setSkipAutoIncrement($noAutoIncrement);
+        ModelMigration::setMigrationPath($migrationsDir);
+
+        $wasMigrated = false;
+        if ($tableName === '@') {
+            $migrations = ModelMigration::generateAll($versionItem, $exportData);
+            foreach ($migrations as $tableName => $migration) {
+                if ($tableName === self::MIGRATION_LOG_TABLE) {
+                    continue;
+                }
+                $tableFile = $migrationPath . DIRECTORY_SEPARATOR . $tableName . '.php';
+                $wasMigrated = file_put_contents(
+                        $tableFile,
+                        '<?php ' . PHP_EOL . PHP_EOL . $migration
+                    ) || $wasMigrated;
+            }
+        } else {
+            $tables = explode(',', $tableName);
+            foreach ($tables as $table) {
+                $migration = ModelMigration::generate($versionItem, $table, $exportData);
+                $tableFile = $migrationPath . DIRECTORY_SEPARATOR . $table . '.php';
+                $wasMigrated = file_put_contents(
+                    $tableFile,
+                    '<?php ' . PHP_EOL . PHP_EOL . $migration
+                );
+            }
+        }
+
+        if (self::isConsole() && $wasMigrated) {
+            print Color::success('Version ' . $versionItem->getVersion() . ' was successfully generated') . PHP_EOL;
+        } elseif (self::isConsole()) {
+            print Color::info('Nothing to generate. You should create tables first.') . PHP_EOL;
+        }
     }
 
     /**
@@ -144,154 +175,400 @@ class Migrations
      * @throws Exception
      * @throws ModelException
      * @throws ScriptException
+     *
      */
     public static function run(array $options)
     {
-        $path = $options['directory'];
-
-        $migrationsDir = $options['migrationsDir'];
-        if (!file_exists($migrationsDir)) {
-            throw new ModelException('Migrations directory could not found.');
+        // Define versioning type to be used
+        if (isset($options['tsBased']) && true === $options['tsBased']) {
+            VersionCollection::setType(VersionCollection::TYPE_TIMESTAMPED);
+        } else {
+            VersionCollection::setType(VersionCollection::TYPE_INCREMENTAL);
         }
 
-        self::$_config = $options['config'];
-        if (!self::$_config instanceof Config) {
-            throw new ModelException('Internal error. Config should be instance of \Phalcon\Config');
+        $migrationsDir = rtrim($options['migrationsDir'], '\\/');
+        if (!file_exists($migrationsDir)) {
+            throw new ModelException('Migrations directory was not found.');
+        }
+
+        /** @var Config $config */
+        $config = $options['config'];
+        if (!$config instanceof Config) {
+            throw new ModelException('Internal error. Config should be an instance of ' . Config::class);
+        }
+
+        // Init ModelMigration
+        if (!isset($config->database)) {
+            throw new ScriptException('Cannot load database configuration');
         }
 
         $finalVersion = null;
         if (isset($options['version']) && $options['version'] !== null) {
-            $finalVersion = new VersionItem($options['version']);
+            $finalVersion = VersionCollection::createItem($options['version']);
         }
 
-        $tableName = 'all';
+        $tableName = '@';
         if (isset($options['tableName'])) {
             $tableName = $options['tableName'];
         }
 
-        $versions = ModelMigration::scanForVersions($migrationsDir);
-        if (!count($versions)) {
-            throw new ModelException("Migrations were not found at {$migrationsDir}");
+        $versionItems = ModelMigration::scanForVersions($migrationsDir);
+
+        if (!isset($versionItems[0])) {
+            throw new ModelException('Migrations were not found at ' . $migrationsDir);
         }
 
-        // set default final version
-        if (!$finalVersion) {
-            $finalVersion = VersionItem::maximum($versions);
+        // Set default final version
+        if ($finalVersion === null) {
+            $finalVersion = VersionCollection::maximum($versionItems);
         }
 
-        // read current version
-        if (is_file($path . '.phalcon')) {
-            unlink($path . '.phalcon');
-            mkdir($path . '.phalcon');
-            chmod($path . '.phalcon', 0775);
-        }
-
-        self::$_migrationFid = $path . '.phalcon/migration-version';
-
-        // init ModelMigration
-        if (!is_null(self::$_config->get('database'))) {
-            throw new ScriptException('Cannot load database configuration');
-        }
-
-        ModelMigration::setup(self::$_config->get('database'));
+        ModelMigration::setup($config->database);
         ModelMigration::setMigrationPath($migrationsDir);
+        self::connectionSetup($options);
+        $initialVersion = self::getCurrentVersion($options);
+        $completedVersions = self::getCompletedVersions($options);
 
-        $initialVersion = self::getCurrentVersion();
-
-        if ($initialVersion->getStamp() == $finalVersion->getStamp()) {
-            return; // nothing to do
+        // Everything is up to date
+        if ($initialVersion->getStamp() === $finalVersion->getStamp()) {
+            print Color::info('Everything is up to date');
+            exit(0);
         }
 
         $direction = ModelMigration::DIRECTION_FORWARD;
         if ($finalVersion->getStamp() < $initialVersion->getStamp()) {
             $direction = ModelMigration::DIRECTION_BACK;
         }
-        // run migration
-        $lastGoodVersion = $initialVersion;
-        $versionsBetween = VersionItem::between($initialVersion, $finalVersion, $versions);
-        foreach ($versionsBetween as $version) {
-            if ($tableName == 'all') {
-                $iterator = new DirectoryIterator($migrationsDir . DIRECTORY_SEPARATOR . $version);
+
+        if (ModelMigration::DIRECTION_FORWARD === $direction) {
+            // If we migrate up, we should go from the beginning to run some migrations which may have been missed
+            $versionItemsTmp = VersionCollection::sortAsc(array_merge($versionItems, [$initialVersion]));
+            $initialVersion = $versionItemsTmp[0];
+        } else {
+            // If we migrate downs, we should go from the last migration to revert some migrations which may have been missed
+            $versionItemsTmp = VersionCollection::sortDesc(array_merge($versionItems, [$initialVersion]));
+            $initialVersion = $versionItemsTmp[0];
+        }
+
+        // Run migration
+        $versionsBetween = VersionCollection::between($initialVersion, $finalVersion, $versionItems);
+        foreach ($versionsBetween as $versionItem) {
+
+            // If we are rolling back, we skip migrating when initialVersion is the same as current
+            if ($initialVersion->getVersion() === $versionItem->getVersion() && ModelMigration::DIRECTION_BACK === $direction) {
+                continue;
+            }
+
+            if ((ModelMigration::DIRECTION_FORWARD === $direction) && isset($completedVersions[(string)$versionItem])) {
+                print Color::info('Version ' . (string)$versionItem . ' was already applied');
+                continue;
+            } elseif ((ModelMigration::DIRECTION_BACK === $direction) && !isset($completedVersions[(string)$initialVersion])) {
+                print Color::info('Version ' . (string)$initialVersion . ' was already rolled back');
+                $initialVersion = $versionItem;
+                continue;
+            }
+
+            if ($initialVersion->getVersion() === $finalVersion->getVersion() && ModelMigration::DIRECTION_BACK === $direction) {
+                break;
+            }
+
+            $migrationStartTime = date("Y-m-d H:i:s");
+            if ($tableName === '@') {
+                // Directory depends on Forward or Back Migration
+                $iterator = new DirectoryIterator(
+                    $migrationsDir . DIRECTORY_SEPARATOR . (ModelMigration::DIRECTION_BACK === $direction ? $initialVersion->getVersion() : $versionItem->getVersion())
+                );
                 foreach ($iterator as $fileInfo) {
                     if (!$fileInfo->isFile() || 0 !== strcasecmp($fileInfo->getExtension(), 'php')) {
                         continue;
                     }
-
-                    ModelMigration::migrate($initialVersion, $version, $fileInfo->getBasename('.php'), $direction);
+                    ModelMigration::migrate($initialVersion, $versionItem, $fileInfo->getBasename('.php'), $direction);
                 }
             } else {
-                ModelMigration::migrate($initialVersion, $version, $tableName, $direction);
+                ModelMigration::migrate($initialVersion, $versionItem, $tableName, $direction);
             }
 
-            self::setCurrentVersion((string)$lastGoodVersion, (string)$version);
-            $lastGoodVersion = $version;
-            print Color::success('Version ' . $version . ' was successfully migrated');
+            if (ModelMigration::DIRECTION_FORWARD == $direction) {
+                self::addCurrentVersion($options, (string)$versionItem, $migrationStartTime);
+                print Color::success('Version ' . $versionItem . ' was successfully migrated');
+            } else {
+                self::removeCurrentVersion($options, (string)$initialVersion);
+                print Color::success('Version ' . $initialVersion . ' was successfully rolled back');
+            }
 
-            $initialVersion = $version;
+            $initialVersion = $versionItem;
         }
     }
 
     /**
-     * @param \Phalcon\Config $database
+     * List migrations along with statuses
+     *
+     * @param array $options
+     *
+     * @throws Exception
+     * @throws ModelException
+     * @throws ScriptException
+     *
+     */
+    public static function listAll(array $options)
+    {
+        // Define versioning type to be used
+        if (true === $options['tsBased']) {
+            VersionCollection::setType(VersionCollection::TYPE_TIMESTAMPED);
+        } else {
+            VersionCollection::setType(VersionCollection::TYPE_INCREMENTAL);
+        }
+
+        $migrationsDir = rtrim($options['migrationsDir'], '/');
+        if (!file_exists($migrationsDir)) {
+            throw new ModelException('Migrations directory was not found.');
+        }
+
+        /** @var Config $config */
+        $config = $options['config'];
+        if (!$config instanceof Config) {
+            throw new ModelException('Internal error. Config should be an instance of ' . Config::class);
+        }
+
+        // Init ModelMigration
+        if (!isset($config->database)) {
+            throw new ScriptException('Cannot load database configuration');
+        }
+
+        $versionItems = ModelMigration::scanForVersions($migrationsDir);
+
+        if (!isset($versionItems[0])) {
+            print Color::info('Migrations were not found at ' . $migrationsDir);
+            return;
+        }
+
+        ModelMigration::setup($config->database);
+        ModelMigration::setMigrationPath($migrationsDir);
+        self::connectionSetup($options);
+
+        $completedVersions = self::getCompletedVersions($options);
+        $versionItems = VersionCollection::sortDesc($versionItems);
+
+        $report = [];
+        foreach ($versionItems as $versionItem) {
+            $versionNumber = $versionItem->getVersion();
+            $report[] = sprintf("│ %-25s │ %12s │", $versionNumber, isset($completedVersions[$versionNumber]) ? 'Y' : 'N');
+        }
+
+        $header = sprintf("│ %-25s │ %12s │", 'Version', 'Was applied');
+        $report[] = '├' . str_repeat('─', 27) . '┼'. str_repeat('─', 14) . '┤';
+        $report[] = $header;
+
+        $report = array_reverse($report);
+
+        echo '┌' . str_repeat('─', 27) . '┬'. str_repeat('─', 14) . '┐' . PHP_EOL;
+        echo join(PHP_EOL, $report) . PHP_EOL;
+        echo '└' . str_repeat('─', 27) . '┴'. str_repeat('─', 14) . '┘'. PHP_EOL . PHP_EOL;
+    }
+
+    /**
+     * Initialize migrations log storage
+     *
+     * @param array $options Applications options
      * @throws DbException
      */
-    public static function connSetup($database)
+    private static function connectionSetup($options)
     {
-        if (!isset($database->adapter)) {
-            throw new DbException('Unspecified database Adapter in your configuration!');
+        if (self::$_storage) {
+            return;
         }
 
-        $adapter = '\\Phalcon\\Db\\Adapter\\Pdo\\' . $database->adapter;
+        if (isset($options['migrationsInDb']) && (bool)$options['migrationsInDb']) {
+            /** @var Config $database */
+            $database = $options['config']['database'];
 
-        if (!class_exists($adapter)) {
-            throw new DbException('Invalid database Adapter!');
-        }
-
-        $configArray = $database->toArray();
-        unset($configArray['adapter']);
-        self::$_connection = new $adapter($configArray);
-
-        if ($database->adapter == 'Mysql') {
-            self::$_connection->query('SET FOREIGN_KEY_CHECKS=0');
-        }
-
-        if (!self::$_connection->tableExists('phalcon_migrations')) {
-            self::$_connection->execute('CREATE TABLE `phalcon_migrations` (`version` VARCHAR(14) NOT NULL, `start_time` TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP, `end_time` TIMESTAMP NOT NULL DEFAULT "0000-00-00 00:00:00" ) DEFAULT CHARSET = utf8;');
-        }
-    }
-
-    public static function getCurrentVersion()
-    {
-        if (!is_null(self::$_config->get('migrationsLog')) && ('database' == self::$_config->get('migrationsLog'))) {
-            if (is_null(self::$_connection)) {
-                self::connSetup(self::$_config->get('database'));
+            if (!isset($database->adapter)) {
+                throw new DbException('Unspecified database Adapter in your configuration!');
             }
 
-            $initialVersion = self::$_connection->query('SELECT `version` FROM `phalcon_migrations` ORDER BY `version` DESC LIMIT 1;');
-            if (0 == $initialVersion->numRows()) {
-                self::$_connection->execute('INSERT INTO `phalcon_migrations` (`version`) VALUES ("0.0.0");');
-                $initialVersion = new VersionItem('0.0.0');
+            $adapter = '\\Phalcon\\Db\\Adapter\\Pdo\\' . $database->adapter;
+
+            if (!class_exists($adapter)) {
+                throw new DbException('Invalid database Adapter!');
+            }
+
+            $configArray = $database->toArray();
+            unset($configArray['adapter']);
+            self::$_storage = new $adapter($configArray);
+
+            if ($database->adapter === 'Mysql') {
+                self::$_storage->query('SET FOREIGN_KEY_CHECKS=0');
+            }
+
+            if (!self::$_storage->tableExists(self::MIGRATION_LOG_TABLE)) {
+                self::$_storage->createTable(self::MIGRATION_LOG_TABLE, null, [
+                    'columns' => [
+                        new Column(
+                            'version',
+                            [
+                                'type' => Column::TYPE_VARCHAR,
+                                'size' => 255,
+                                'notNull' => true,
+                            ]
+                        ),
+                        new Column(
+                            'start_time',
+                            [
+                                'type' => Column::TYPE_TIMESTAMP,
+                                'notNull' => true,
+                                'default' => 'CURRENT_TIMESTAMP',
+                            ]
+                        ),
+                        new Column(
+                            'end_time',
+                            [
+                                'type' => 'TIMESTAMP NOT NULL DEFAULT NOW()',
+                            ]
+                        )
+                    ],
+                    'indexes' => [
+                        new Index('idx_' . self::MIGRATION_LOG_TABLE . '_version', ['version'])
+                    ]
+                ]);
+            }
+
+        } else {
+            if (empty($options['directory'])) {
+                $path = defined('BASE_PATH') ? BASE_PATH : defined('APP_PATH') ? dirname(APP_PATH) : '';
+                $path = rtrim($path, '\\/') . '/.phalcon';
             } else {
-                $initialVersion = $initialVersion->fetch();
-                $initialVersion = new VersionItem($initialVersion['version']);
+                $path = rtrim($options['directory'], '\\/') . '/.phalcon';
             }
-        } else {
-            $initialVersion = new VersionItem(file_exists(self::$_migrationFid) ? file_get_contents(self::$_migrationFid) : null);
-        }
+            if (!is_dir($path) && !is_writable(dirname($path))) {
+                throw new \RuntimeException("Unable to write '{$path}' directory. Permission denied");
+            }
+            if (is_file($path)) {
+                unlink($path);
+                mkdir($path);
+                chmod($path, 0775);
+            } elseif (!is_dir($path)) {
+                mkdir($path);
+                chmod($path, 0775);
+            }
 
-        return $initialVersion;
+            self::$_storage = $path . '/migration-version';
+
+            if (!file_exists(self::$_storage)) {
+                if (!is_writable($path)) {
+                    throw new \RuntimeException("Unable to write '" . self::$_storage . "' file. Permission denied");
+                }
+                touch(self::$_storage);
+            }
+        }
     }
 
-    public static function setCurrentVersion($lastVersion, $currentVersion)
+    /**
+     * Get latest completed migration version
+     *
+     * @param array $options Applications options
+     * @return ItemInterface
+     */
+    public static function getCurrentVersion($options)
     {
-        if (is_null(self::$_connection)) {
-            self::connSetup(self::$_config->get('database'));
+        self::connectionSetup($options);
+
+        if (isset($options['migrationsInDb']) && (bool)$options['migrationsInDb']) {
+            /** @var AdapterInterface $connection */
+            $connection = self::$_storage;
+            $lastGoodMigration = $connection->query('SELECT * FROM '. self::MIGRATION_LOG_TABLE .' ORDER BY version DESC LIMIT 1');
+            if (0 == $lastGoodMigration->numRows()) {
+                return VersionCollection::createItem(null);
+            } else {
+                $lastGoodMigration = $lastGoodMigration->fetchArray();
+
+                return VersionCollection::createItem($lastGoodMigration['version']);
+            }
+        } else {
+            // Get and clean migration
+            $version = file_exists(self::$_storage) ? file_get_contents(self::$_storage) : null;
+
+            if ($version = trim($version) ?: null) {
+                $version = preg_split('/\r\n|\r|\n/', $version, -1, PREG_SPLIT_NO_EMPTY);
+                $version = array_pop($version);
+            }
+
+            return VersionCollection::createItem($version);
+        }
+    }
+
+    /**
+     * Add migration version to log
+     *
+     * @param array $options Applications options
+     * @param string $version Migration version to store
+     * @param string $startTime Migration start timestamp
+     */
+    public static function addCurrentVersion($options, $version, $startTime = null)
+    {
+        self::connectionSetup($options);
+
+        if ($startTime === null) {
+            $startTime = date("Y-m-d H:i:s");
+        }
+        $endTime = date("Y-m-d H:i:s");
+
+        if (isset($options['migrationsInDb']) && (bool) $options['migrationsInDb']) {
+            /** @var AdapterInterface $connection */
+            $connection = self::$_storage;
+            $connection->insert(self::MIGRATION_LOG_TABLE, [$version, $startTime, $endTime], ['version', 'start_time', 'end_time']);
+        } else {
+            $currentVersions = self::getCompletedVersions($options);
+            $currentVersions[(string)$version] = 1;
+            $currentVersions = array_keys($currentVersions);
+            sort($currentVersions);
+            file_put_contents(self::$_storage, implode("\n", $currentVersions));
+        }
+    }
+
+    /**
+     * Remove migration version from log
+     *
+     * @param array $options Applications options
+     * @param string $version Migration version to remove
+     */
+    public static function removeCurrentVersion($options, $version)
+    {
+        self::connectionSetup($options);
+
+        if (isset($options['migrationsInDb']) && (bool)$options['migrationsInDb']) {
+            /** @var AdapterInterface $connection */
+            $connection = self::$_storage;
+            $connection->execute('DELETE FROM '. self::MIGRATION_LOG_TABLE .' WHERE version=\'' . $version . '\' LIMIT 1');
+        } else {
+            $currentVersions = self::getCompletedVersions($options);
+            unset($currentVersions[(string)$version]);
+            $currentVersions = array_keys($currentVersions);
+            sort($currentVersions);
+            file_put_contents(self::$_storage, implode("\n", $currentVersions));
+        }
+    }
+
+    /**
+     * Scan $_storage for all completed versions
+     *
+     * @param array $options Applications options
+     * @return array
+     */
+    public static function getCompletedVersions($options)
+    {
+        self::connectionSetup($options);
+
+        if (isset($options['migrationsInDb']) && (bool)$options['migrationsInDb']) {
+            /** @var AdapterInterface $connection */
+            $connection = self::$_storage;
+            $completedVersions = $connection->query('SELECT version FROM '. self::MIGRATION_LOG_TABLE .' ORDER BY version DESC')->fetchAll();
+            $completedVersions = array_map(function ($version) {
+                return $version['version'];
+            }, $completedVersions);
+        } else {
+            $completedVersions = file(self::$_storage, FILE_IGNORE_NEW_LINES);
         }
 
-        if (!is_null(self::$_config->get('migrationsLog')) && ('database' == self::$_config->get('migrationsLog'))) {
-            self::$_connection->execute('UPDATE `phalcon_migrations` SET `version`="' . (string)$currentVersion . '" WHERE `version`="' . (string)$lastVersion . '" LIMIT 1;');
-        } else {
-            file_put_contents(self::$_migrationFid, (string)$currentVersion);
-        }
+        return array_flip($completedVersions);
     }
 }
