@@ -4,10 +4,10 @@
   +------------------------------------------------------------------------+
   | Phalcon Developer Tools                                                |
   +------------------------------------------------------------------------+
-  | Copyright (c) 2011-2016 Phalcon Team (http://www.phalconphp.com)       |
+  | Copyright (c) 2011-2016 Phalcon Team (https://www.phalconphp.com)      |
   +------------------------------------------------------------------------+
   | This source file is subject to the New BSD License that is bundled     |
-  | with this package in the file docs/LICENSE.txt.                        |
+  | with this package in the file LICENSE.txt.                             |
   |                                                                        |
   | If you did not receive a copy of the license and are unable to         |
   | obtain it through the world-wide-web, please send an email             |
@@ -21,12 +21,18 @@
 
 namespace Phalcon;
 
-use Phalcon\Config;
-
 class Utils
 {
     const DB_ADAPTER_POSTGRESQL = 'Postgresql';
 
+    const DB_ADAPTER_SQLITE = 'Sqlite';
+
+    /**
+     * Converts the underscore_notation to the UpperCamelCase
+     *
+     * @param string $string
+     * @return string
+     */
     public static function camelize($string)
     {
         $stringParts = explode('_', $string);
@@ -36,9 +42,20 @@ class Utils
     }
 
     /**
+     * Converts the underscore_notation to the lowerCamelCase
+     *
+     * @param string $string
+     * @return string
+     */
+    public static function lowerCamelize($string)
+    {
+        return lcfirst(self::camelize($string));
+    }
+
+    /**
      * Resolves the DB Schema
      *
-     * @param \Phalcon\Config $config
+     * @param Config $config
      * @return null|string
      */
     public static function resolveDbSchema(Config $config)
@@ -48,7 +65,13 @@ class Utils
         }
 
         if (self::DB_ADAPTER_POSTGRESQL == $config->get('adapter')) {
-            return  'public';
+            return 'public';
+        }
+
+        if (self::DB_ADAPTER_SQLITE == $config->get('adapter')) {
+            // SQLite only supports the current database, unless one is
+            // attached. This is not the case, so don't return a schema.
+            return null;
         }
 
         if ($config->offsetExists('dbname')) {
