@@ -5,6 +5,7 @@ namespace @@FQMN@@;
 use Phalcon\DiInterface;
 use Phalcon\Loader;
 use Phalcon\Mvc\View;
+use Phalcon\Mvc\View\Engine\Php as PhpEngine;
 use Phalcon\Mvc\ModuleDefinitionInterface;
 @@iniConfigImport@@
 @@useConfig@@
@@ -20,10 +21,10 @@ class Module implements ModuleDefinitionInterface
     {
         $loader = new Loader();
 
-        $loader->registerNamespaces(array(
+        $loader->registerNamespaces([
             '@@FQMN@@\Controllers' => __DIR__ . '/controllers/',
             '@@FQMN@@\Models'      => __DIR__ . '/models/'
-        ));
+        ]);
 
         $loader->register();
     }
@@ -36,14 +37,12 @@ class Module implements ModuleDefinitionInterface
     public function registerServices(DiInterface $di)
     {
         /**
-         * Read common configuration
-         */
-         $config = $di->has('config') ? $di->getShared('config') : null;
-
-        /**
          * Try to load local configuration
          */
         if (file_exists(@@configName@@)) {
+            
+            $config = $di['config'];
+            
             $override = @@configLoader@@;
 
             if ($config instanceof Config) {
@@ -56,9 +55,16 @@ class Module implements ModuleDefinitionInterface
         /**
          * Setting up the view component
          */
-        $di['view'] = function () use ($config) {
+        $di['view'] = function () {
+            $config = $this->getConfig();
+
             $view = new View();
             $view->setViewsDir($config->get('application')->viewsDir);
+            
+            $view->registerEngines([
+                '.volt'  => 'voltShared',
+                '.phtml' => PhpEngine::class
+            ]);
 
             return $view;
         };
@@ -66,13 +72,15 @@ class Module implements ModuleDefinitionInterface
         /**
          * Database connection is created based in the parameters defined in the configuration file
          */
-        $di['db'] = function () use ($config) {
-            $config = $config->database->toArray();
+        $di['db'] = function () {
+            $config = $this->getConfig();
 
-            $dbAdapter = '\Phalcon\Db\Adapter\Pdo\\' . $config['adapter'];
+            $dbConfig = $config->database->toArray();
+
+            $dbAdapter = '\Phalcon\Db\Adapter\Pdo\\' . $dbConfig['adapter'];
             unset($config['adapter']);
 
-            return new $dbAdapter($config);
+            return new $dbAdapter($dbConfig);
         };
     }
 }
