@@ -1,9 +1,9 @@
 <?php
 
-return array(
-    'application' => array(
+return [
+    'application' => [
         'controllersDir' => __DIR__ . '/../controllers/',
         'modelsDir'      => __DIR__ . '/../models/',
         'viewsDir'       => __DIR__ . '/../views/',
-    )
-);
+    ]
+];
