@@ -1,7 +1,14 @@
 <?php
 
-return new \Phalcon\Config(array(
-
+return new \Phalcon\Config([
+    'database' => [
+      'adapter'     => 'Mysql',
+      'host'        => 'localhost',
+      'username'    => 'root',
+      'password'    => '',
+      'dbname'      => '@@name@@',
+      'charset'     => 'utf8',
+    ],
     'version' => '1.0',
 
     /**
@@ -14,4 +21,4 @@ return new \Phalcon\Config(array(
      * You can disable this behaviour if the output of your application needs to don't have a new line at end
      */
     'printNewLine' => true
-));
+]);
