@@ -1,11 +1,34 @@
 <?php
-/**
- * Local variables
- *
- * @var \Phalcon\Config $config
- * @var \Phalcon\Di\FactoryDefault\Cli $di
- */
 
-$di->setShared('config', function () use ($config) {
-    return $config;
+/**
+ * Shared configuration service
+ */
+$di->setShared('config', function () {
+    return include APP_PATH . '/config/config.php';
 });
+
+/**
+ * Database connection is created based in the parameters defined in the configuration file
+ */
+$di->setShared('db', function () {
+    $config = $this->getConfig();
+
+    $class = 'Phalcon\Db\Adapter\Pdo\\' . $config->database->adapter;
+
+    $params = [
+        'host'     => $config->database->host,
+        'username' => $config->database->username,
+        'password' => $config->database->password,
+        'dbname'   => $config->database->dbname,
+        'charset'  => $config->database->charset
+    ];
+
+    if ($config->database->adapter == 'Postgresql') {
+        unset($params['charset']);
+    }
+
+    $connection = new $class($params);
+
+    return $connection;
+});
+
