@@ -7,14 +7,14 @@
 /**
  * Add your routes here
  */
-$app->get('/', function () use ($app) {
-    echo $app['view']->render('index');
+$app->get('/', function () {
+    echo $this['view']->render('index');
 });
 
 /**
  * Not found handler
  */
-$app->notFound(function () use ($app) {
+$app->notFound(function () use($app) {
     $app->response->setStatusCode(404, "Not Found")->sendHeaders();
     echo $app['view']->render('404');
 });
