@@ -1,22 +1,25 @@
 <?php
+/*
+ * Modified: prepend directory path of current file, because of this file own different ENV under between Apache and command line.
+ * NOTE: please remove this comment.
+ */
+defined('BASE_PATH') || define('BASE_PATH', getenv('BASE_PATH') ?: realpath(dirname(__FILE__) . '/../..'));
+defined('APP_PATH') || define('APP_PATH', BASE_PATH . '/app');
 
-defined('APP_PATH') || define('APP_PATH', realpath('.'));
-
-return new \Phalcon\Config(array(
-
-    'database' => array(
+return new \Phalcon\Config([
+    'database' => [
         'adapter'    => 'Mysql',
         'host'       => 'localhost',
         'username'   => 'root',
         'password'   => '',
         'dbname'     => 'test',
         'charset'    => 'utf8',
-    ),
+    ],
 
-    'application' => array(
+    'application' => [
         'modelsDir'      => APP_PATH . '/models/',
         'migrationsDir'  => APP_PATH . '/migrations/',
         'viewsDir'       => APP_PATH . '/views/',
         'baseUri'        => '/@@name@@/',
-    )
-));
+    ]
+]);
