@@ -1,22 +1,30 @@
 <?php
 
+use Phalcon\Di\FactoryDefault;
 use Phalcon\Mvc\Micro;
 
 error_reporting(E_ALL);
 
-define('APP_PATH', realpath('..'));
+define('BASE_PATH', dirname(__DIR__));
+define('APP_PATH', BASE_PATH . '/app');
 
 try {
 
     /**
-     * Read the configuration
+     * The FactoryDefault Dependency Injector automatically registers the services that
+     * provide a full stack framework. These default services can be overidden with custom ones.
      */
-    $config = @@configLoader@@;
+    $di = new FactoryDefault();
 
     /**
      * Include Services
      */
     include APP_PATH . '/config/services.php';
+
+    /**
+     * Get config service for use in inline setup below
+     */
+    $config = $di->getConfig();
 
     /**
      * Include Autoloader
