@@ -6,7 +6,7 @@
 $loader = new \Phalcon\Loader();
 
 $loader->registerDirs(
-    array(
+    [
         $config->application->modelsDir
-    )
+    ]
 )->register();
