@@ -1,20 +1,21 @@
 <?php
-/**
- * Services are globally registered in this file
- *
- * @var \Phalcon\Config $config
- */
 
 use Phalcon\Mvc\View\Simple as View;
 use Phalcon\Mvc\Url as UrlResolver;
-use Phalcon\Di\FactoryDefault;
 
-$di = new FactoryDefault();
+/**
+ * Shared configuration service
+ */
+$di->setShared('config', function () {
+    return @@configLoader@@;
+});
 
 /**
  * Sets the view component
  */
-$di->setShared('view', function () use ($config) {
+$di->setShared('view', function () {
+    $config = $this->getConfig();
+
     $view = new View();
     $view->setViewsDir($config->application->viewsDir);
     return $view;
@@ -23,7 +24,9 @@ $di->setShared('view', function () use ($config) {
 /**
  * The URL component is used to generate all kind of urls in the application
  */
-$di->setShared('url', function () use ($config) {
+$di->setShared('url', function () {
+    $config = $this->getConfig();
+
     $url = new UrlResolver();
     $url->setBaseUri($config->application->baseUri);
     return $url;
@@ -32,12 +35,24 @@ $di->setShared('url', function () use ($config) {
 /**
  * Database connection is created based in the parameters defined in the configuration file
  */
-$di->setShared('db', function () use ($config) {
-    $dbConfig = $config->database->toArray();
-    $adapter = $dbConfig['adapter'];
-    unset($dbConfig['adapter']);
+$di->setShared('db', function () {
+    $config = $this->getConfig();
 
-    $class = 'Phalcon\Db\Adapter\Pdo\\' . $adapter;
+    $class = 'Phalcon\Db\Adapter\Pdo\\' . $config->database->adapter;
+    $params = [
+        'host'     => $config->database->host,
+        'username' => $config->database->username,
+        'password' => $config->database->password,
+        'dbname'   => $config->database->dbname,
+        'charset'  => $config->database->charset
+    ];
 
-    return new $class($dbConfig);
+    if ($config->database->adapter == 'Postgresql') {
+        unset($params['charset']);
+    }
+
+    $connection = new $class($params);
+
+    return $connection;
 });
+
