@@ -1,30 +1,28 @@
 <?php
 
-$router = $di->get("router");
+$router = $di->getRouter();
 
 foreach ($application->getModules() as $key => $module) {
-    $namespace = str_replace('Module','Controllers', $module["className"]);
-    $router->add('/'.$key.'/:params', array(
+    $namespace = preg_replace('/Module$/', 'Controllers', $module["className"]);
+    $router->add('/'.$key.'/:params', [
         'namespace' => $namespace,
         'module' => $key,
         'controller' => 'index',
         'action' => 'index',
         'params' => 1
-    ))->setName($key);
-    $router->add('/'.$key.'/:controller/:params', array(
+    ])->setName($key);
+    $router->add('/'.$key.'/:controller/:params', [
         'namespace' => $namespace,
         'module' => $key,
         'controller' => 1,
         'action' => 'index',
         'params' => 2
-    ));
-    $router->add('/'.$key.'/:controller/:action/:params', array(
+    ]);
+    $router->add('/'.$key.'/:controller/:action/:params', [
         'namespace' => $namespace,
         'module' => $key,
         'controller' => 1,
         'action' => 2,
         'params' => 3
-    ));
+    ]);
 }
-
-$di->set("router", $router);
