@@ -1,84 +1,39 @@
 <?php
-/**
- * Services are globally registered in this file
- *
- * @var \Phalcon\Config $config
- */
 
-use Phalcon\Mvc\Router;
-use Phalcon\Mvc\Url as UrlResolver;
-use Phalcon\Di\FactoryDefault;
-use Phalcon\Session\Adapter\Files as SessionAdapter;
+use Phalcon\Loader;
 use Phalcon\Mvc\Model\Metadata\Memory as MetaDataAdapter;
-use Phalcon\Mvc\View;
 use Phalcon\Mvc\View\Engine\Volt as VoltEngine;
-use Phalcon\Flash\Direct as Flash;
+
 
 /**
- * The FactoryDefault Dependency Injector automatically registers the right services to provide a full stack framework
+ * Shared configuration service
  */
-$di = new FactoryDefault();
-
-/**
- * Registering a router
- */
-$di->setShared('router', function () {
-    $router = new Router();
-
-    $router->setDefaultModule('frontend');
-    $router->setDefaultNamespace('@@namespace@@\Frontend\Controllers');
-
-    return $router;
-});
-
-/**
- * The URL component is used to generate all kinds of URLs in the application
- */
-$di->setShared('url', function () use ($config) {
-    $url = new UrlResolver();
-    $url->setBaseUri($config->application->baseUri);
-
-    return $url;
-});
-
-/**
- * Setting up the view component
- */
-$di->setShared('view', function () use ($config) {
-
-    $view = new View();
-
-    $view->setViewsDir($config->application->viewsDir);
-
-    $view->registerEngines(array(
-        '.volt' => function ($view, $di) use ($config) {
-
-            $volt = new VoltEngine($view, $di);
-
-            $volt->setOptions(array(
-                'compiledPath' => $config->application->cacheDir,
-                'compiledSeparator' => '_'
-            ));
-
-            return $volt;
-        },
-        '.phtml' => 'Phalcon\Mvc\View\Engine\Php'
-    ));
-
-    return $view;
+$di->setShared('config', function () {
+    return @@configLoader@@;
 });
 
 /**
  * Database connection is created based in the parameters defined in the configuration file
  */
-$di->setShared('db', function () use ($config) {
-    $dbConfig = $config->database->toArray();
-    $adapter = $dbConfig['adapter'];
-    unset($dbConfig['adapter']);
+$di->setShared('db', function () {
+    $config = $this->getConfig();
 
-    $class = 'Phalcon\Db\Adapter\Pdo\\' . $adapter;
+    $class = 'Phalcon\Db\Adapter\Pdo\\' . $config->database->adapter;
+    $params = [
+        'host'     => $config->database->host,
+        'username' => $config->database->username,
+        'password' => $config->database->password,
+        'dbname'   => $config->database->dbname,
+        'charset'  => $config->database->charset
+    ];
 
-    return new $class($dbConfig);
+    if ($config->database->adapter == 'Postgresql') {
+        unset($params['charset']);
+    }
+
+    $connection = new $class($params);
+
+    return $connection;
 });
 
 /**
@@ -89,32 +44,24 @@ $di->setShared('modelsMetadata', function () {
 });
 
 /**
- * Starts the session the first time some component requests the session service
+ * Configure the Volt service for rendering .volt templates
  */
-$di->setShared('session', function () {
-    $session = new SessionAdapter();
-    $session->start();
+$di->setShared('voltShared', function ($view) {
+    $config = $this->getConfig();
 
-    return $session;
-});
+    $volt = new VoltEngine($view, $this);
+    $volt->setOptions([
+        'compiledPath' => function($templatePath) use ($config) {
 
-/**
- * Register the session flash service with the Twitter Bootstrap classes
- */
-$di->set('flash', function () {
-    return new Flash(array(
-        'error'   => 'alert alert-danger',
-        'success' => 'alert alert-success',
-        'notice'  => 'alert alert-info',
-        'warning' => 'alert alert-warning'
-    ));
-});
+            // Makes the view path into a portable fragment
+            $templateFrag = str_replace($config->application->appDir, '', $templatePath);
 
-/**
-* Set the default namespace for dispatcher
-*/
-$di->setShared('dispatcher', function() use ($di) {
-    $dispatcher = new Phalcon\Mvc\Dispatcher();
-    $dispatcher->setDefaultNamespace('@@namespace@@\Frontend\Controllers');
-    return $dispatcher;
+            // Replace '/' with a safe '%%'
+            $templateFrag = str_replace('/', '%%', $templateFrag);
+
+            return $config->application->cacheDir . 'volt/' . $templateFrag . '.php';
+        }
+    ]);
+
+    return $volt;
 });
