@@ -1,32 +1,45 @@
 <?php
+use Phalcon\Di\FactoryDefault;
 
 error_reporting(E_ALL);
 
-define('APP_PATH', realpath('..'));
+define('BASE_PATH', dirname(__DIR__));
+define('APP_PATH', BASE_PATH . '/app');
 
 try {
 
     /**
-     * Read the configuration
+     * The FactoryDefault Dependency Injector automatically registers
+     * the services that provide a full stack framework.
      */
-    $config = @@configLoader@@;
+    $di = new FactoryDefault();
 
     /**
-     * Read auto-loader
+     * Handle routes
      */
-    include APP_PATH . "/app/config/loader.php";
+    include APP_PATH . '/config/router.php';
 
     /**
      * Read services
      */
-    include APP_PATH . "/app/config/services.php";
+    include APP_PATH . '/config/services.php';
+
+    /**
+     * Get config service for use in inline setup below
+     */
+    $config = $di->getConfig();
+
+    /**
+     * Include Autoloader
+     */
+    include APP_PATH . '/config/loader.php';
 
     /**
      * Handle the request
      */
     $application = new \Phalcon\Mvc\Application($di);
 
-    echo $application->handle()->getContent();
+    echo str_replace(["\n","\r","\t"], '', $application->handle()->getContent());
 
 } catch (\Exception $e) {
     echo $e->getMessage() . '<br>';
