@@ -1,12 +1,7 @@
 <?php
-/**
- * Services are globally registered in this file
- *
- * @var \Phalcon\Config $config
- */
 
-use Phalcon\Di\FactoryDefault;
 use Phalcon\Mvc\View;
+use Phalcon\Mvc\View\Engine\Php as PhpEngine;
 use Phalcon\Mvc\Url as UrlResolver;
 use Phalcon\Mvc\View\Engine\Volt as VoltEngine;
 use Phalcon\Mvc\Model\Metadata\Memory as MetaDataAdapter;
@@ -14,14 +9,18 @@ use Phalcon\Session\Adapter\Files as SessionAdapter;
 use Phalcon\Flash\Direct as Flash;
 
 /**
- * The FactoryDefault Dependency Injector automatically register the right services providing a full stack framework
+ * Shared configuration service
  */
-$di = new FactoryDefault();
+$di->setShared('config', function () {
+    return @@configLoader@@;
+});
 
 /**
  * The URL component is used to generate all kind of urls in the application
  */
-$di->setShared('url', function () use ($config) {
+$di->setShared('url', function () {
+    $config = $this->getConfig();
+
     $url = new UrlResolver();
     $url->setBaseUri($config->application->baseUri);
 
@@ -31,26 +30,29 @@ $di->setShared('url', function () use ($config) {
 /**
  * Setting up the view component
  */
-$di->setShared('view', function () use ($config) {
+$di->setShared('view', function () {
+    $config = $this->getConfig();
 
     $view = new View();
-
+    $view->setDI($this);
     $view->setViewsDir($config->application->viewsDir);
 
-    $view->registerEngines(array(
-        '.volt' => function ($view, $di) use ($config) {
+    $view->registerEngines([
+        '.volt' => function ($view) {
+            $config = $this->getConfig();
 
-            $volt = new VoltEngine($view, $di);
+            $volt = new VoltEngine($view, $this);
 
-            $volt->setOptions(array(
+            $volt->setOptions([
                 'compiledPath' => $config->application->cacheDir,
                 'compiledSeparator' => '_'
-            ));
+            ]);
 
             return $volt;
         },
-        '.phtml' => 'Phalcon\Mvc\View\Engine\Php'
-    ));
+        '.phtml' => PhpEngine::class
+
+    ]);
 
     return $view;
 });
@@ -58,15 +60,27 @@ $di->setShared('view', function () use ($config) {
 /**
  * Database connection is created based in the parameters defined in the configuration file
  */
-$di->setShared('db', function () use ($config) {
-    $dbConfig = $config->database->toArray();
-    $adapter = $dbConfig['adapter'];
-    unset($dbConfig['adapter']);
+$di->setShared('db', function () {
+    $config = $this->getConfig();
 
-    $class = 'Phalcon\Db\Adapter\Pdo\\' . $adapter;
+    $class = 'Phalcon\Db\Adapter\Pdo\\' . $config->database->adapter;
+    $params = [
+        'host'     => $config->database->host,
+        'username' => $config->database->username,
+        'password' => $config->database->password,
+        'dbname'   => $config->database->dbname,
+        'charset'  => $config->database->charset
+    ];
 
-    return new $class($dbConfig);
+    if ($config->database->adapter == 'Postgresql') {
+        unset($params['charset']);
+    }
+
+    $connection = new $class($params);
+
+    return $connection;
 });
+
 
 /**
  * If the configuration specify the use of metadata adapter use it or use memory otherwise
@@ -79,12 +93,12 @@ $di->setShared('modelsMetadata', function () {
  * Register the session flash service with the Twitter Bootstrap classes
  */
 $di->set('flash', function () {
-    return new Flash(array(
+    return new Flash([
         'error'   => 'alert alert-danger',
         'success' => 'alert alert-success',
         'notice'  => 'alert alert-info',
         'warning' => 'alert alert-warning'
-    ));
+    ]);
 });
 
 /**
