@@ -4,10 +4,10 @@
   +------------------------------------------------------------------------+
   | Phalcon Developer Tools                                                |
   +------------------------------------------------------------------------+
-  | Copyright (c) 2011-2015 Phalcon Team (http://www.phalconphp.com)       |
+  | Copyright (c) 2011-2016 Phalcon Team (https://www.phalconphp.com)      |
   +------------------------------------------------------------------------+
   | This source file is subject to the New BSD License that is bundled     |
-  | with this package in the file docs/LICENSE.txt.                        |
+  | with this package in the file LICENSE.txt.                             |
   |                                                                        |
   | If you did not receive a copy of the license and are unable to         |
   | obtain it through the world-wide-web, please send an email             |
@@ -15,6 +15,7 @@
   +------------------------------------------------------------------------+
   | Authors: Andres Gutierrez <andres@phalconphp.com>                      |
   |          Eduar Carvajal <eduar@phalconphp.com>                         |
+  |          Serghei Iakovlev <serghei@phalconphp.com>                     |
   +------------------------------------------------------------------------+
 */
 
@@ -27,6 +28,33 @@
  *          192.168.0.1 or SUBNET 192., 10.0.2., 86.84.124.
  */
 defined('PTOOLS_IP') || define('PTOOLS_IP', '192.168.');
+defined('BASE_PATH') || define('BASE_PATH', dirname(dirname(__FILE__)));
+defined('APP_PATH') || define('APP_PATH', BASE_PATH . DIRECTORY_SEPARATOR . 'app');
+
+/**
+ * @const ENV_PRODUCTION Application production stage.
+ */
+defined('ENV_PRODUCTION') || define('ENV_PRODUCTION', 'production');
+
+/**
+ * @const ENV_STAGING Application staging stage.
+ */
+defined('ENV_STAGING') || define('ENV_STAGING', 'staging');
+
+/**
+ * @const ENV_DEVELOPMENT Application development stage.
+ */
+defined('ENV_DEVELOPMENT') || define('ENV_DEVELOPMENT', 'development');
+
+/**
+ * @const ENV_TESTING Application test stage.
+ */
+defined('ENV_TESTING') || define('ENV_TESTING', 'testing');
+
+/**
+ * @const APPLICATION_ENV Current application stage.
+ */
+defined('APPLICATION_ENV') || define('APPLICATION_ENV', getenv('APPLICATION_ENV') ?: ENV_DEVELOPMENT);
 
 // ---------------------------- DO NOT EDIT BELOW ------------------------------
 
